requirements.txt
acme==1.12.0
attrs==20.3.0
authres==1.2.0
Automat==20.2.0
bcrypt==3.1.7
certbot==1.12.0
certbot-nginx==1.10.1
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
ConfigArgParse==1.2.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.3.2
cysystemd==1.5.4
dbus-python==1.2.16
deluge==2.0.3
distro==1.5.0
dnspython==2.3.0
GeoIP==1.3.2
gpiozero==1.6.2
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
josepy==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.5
parsedatetime==2.6
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.38.0
PyHamcrest==1.9.0
PyICU==2.5
pymongo==4.3.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-libtorrent==1.2.9
python-prctl==1.7
pytz==2021.1
pyxdg==0.27
rencode==1.0.6
requests==2.25.1
requests-toolbelt==0.9.1
RPi.GPIO==0.7.0
schedule==1.2.0
service-identity==18.1.0
setproctitle==1.2.1
simplejpeg==1.6.4
six==1.16.0
spf-engine==2.9.2
spidev==3.5
ssh-import-id==5.10
systemd==0.17.1
toml==0.10.1
Twisted==20.3.0
ufw==0.36
urllib3==1.26.5
v4l2-python3==0.3.2
xdg==5
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
